propoto/api/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
starlette>=0.45.0

# AI & Agents
pydantic-ai-slim>=1.25.0
openai>=1.0.0

# Environment & Config
python-dotenv==1.2.1

# HTTP Client
httpx==0.28.1

# External Service Integrations
firecrawl-py==4.9.0
# Removed: mem0ai (not in use - Exa and Firecrawl only)

# Telegram Bot
python-telegram-bot==21.7

# Retry Logic
tenacity==9.0.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==4.1.0
pytest-mock==3.12.0
